requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
abseil-cpp=20210324.2=h23ab428_0
absl-py=0.15.0=pyhd3eb1b0_0
aggdraw=1.3.14=pypi_0
aiohttp=3.8.1=py38hca72f7f_0
aiosignal=1.2.0=pyhd3eb1b0_0
aom=3.2.0=he49afe7_2
appnope=0.1.2=py38hecd8cb5_1001
argon2-cffi=20.1.0=py38h9ed2024_1
astunparse=1.6.3=py_0
async-timeout=4.0.1=pyhd3eb1b0_0
async_generator=1.10=pyhd3eb1b0_0
attrs=21.2.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
bleach=4.1.0=pyhd3eb1b0_0
blinker=1.4=py38hecd8cb5_0
bottleneck=1.3.2=py38hf1fa96c_1
brotli=1.0.9=hb1e8313_2
brotlipy=0.7.0=py38h9ed2024_1003
bzip2=1.0.8=h1de35cc_0
c-ares=1.18.1=hca72f7f_0
ca-certificates=2020.10.14=0
cachetools=4.2.2=pyhd3eb1b0_0
cairo=1.16.0=h8023c5d_1
certifi=2020.6.20=py38_0
cffi=1.15.0=py38h1a44b6c_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.3=pyhd3eb1b0_0
cloudpickle=2.0.0=pyhd3eb1b0_0
cryptography=3.4.8=py38h2fd3fbb_0
cycler=0.11.0=pyhd3eb1b0_0
cytoolz=0.11.0=py38haf1e3a3_0
dask-core=2021.10.0=pyhd3eb1b0_0
dataclasses=0.8=pyh6d0b6a4_7
debugpy=1.5.1=py38he9d5cce_0
decorator=5.1.0=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
dill=0.3.4=pyhd3eb1b0_0
entrypoints=0.3=py38_0
ffmpeg=4.4.1=h79e7b16_0
fontconfig=2.13.1=ha9ee91d_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=hd8bbffd_0
frozenlist=1.2.0=py38hca72f7f_0
fsspec=2021.10.1=pyhd3eb1b0_0
future=0.18.2=py38_1
gast=0.3.3=py_0
gettext=0.21.0=h7535e17_0
giflib=5.2.1=haf1e3a3_0
glib=2.70.2=hcf210ce_1
glib-tools=2.70.2=hcf210ce_1
gmp=6.2.1=h23ab428_2
gnutls=3.6.15=hed9c0bf_0
google-auth=1.33.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
google-pasta=0.2.0=pyhd3eb1b0_0
googleapis-common-protos=1.53.0=py38hecd8cb5_0
graphite2=1.3.14=h38d11af_0
grpc-cpp=1.37.1=h69eed7d_3
grpcio=1.37.1=py38ha263829_0
h5py=2.10.0=py38h0601b69_1
harfbuzz=2.9.1=h159f659_1
hdf5=1.10.6=nompi_hc5d9132_1114
icu=68.1=h23ab428_0
idna=3.3=pyhd3eb1b0_0
imageio=2.9.0=pyhd3eb1b0_0
importlib-metadata=4.8.2=py38hecd8cb5_0
importlib_metadata=4.8.2=hd3eb1b0_0
iniconfig=1.1.1=pyhd3eb1b0_0
ipykernel=6.4.1=py38hecd8cb5_1
ipython=7.29.0=py38h01d92e1_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jasper=1.900.1=h1f36771_4
jedi=0.18.0=py38hecd8cb5_1
jinja2=3.0.2=pyhd3eb1b0_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9d=h9ed2024_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter_client=7.1.0=pyhd3eb1b0_0
jupyter_core=4.9.1=py38hecd8cb5_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
keras=2.4.3=hd3eb1b0_0
keras-base=2.4.3=pyhd3eb1b0_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.3.1=py38h23ab428_0
krb5=1.19.2=hcd88c3b_0
lame=3.100=h1de35cc_0
lcms2=2.12=hf1fd2bf_0
libblas=3.9.0=12_osx64_openblas
libcblas=3.9.0=12_osx64_openblas
libcurl=7.80.0=h6dfd666_0
libcxx=12.0.0=h2f01273_0
libedit=3.1.20210910=hca72f7f_0
libev=4.33=h9ed2024_1
libffi=3.4.2=he9d5cce_2
libgfortran=5.0.0=9_3_0_h6c81a4c_23
libgfortran5=9.3.0=h6c81a4c_23
libglib=2.70.2=hf1fb8c0_1
libiconv=1.16=h1de35cc_0
libidn2=2.3.2=h9ed2024_0
liblapack=3.9.0=12_osx64_openblas
liblapacke=3.9.0=12_osx64_openblas
libnghttp2=1.46.0=ha29bfda_0
libopenblas=0.3.18=openmp_h3351f45_0
libopencv=4.5.1=py38h229d5d6_1
libpng=1.6.37=ha441bb4_0
libprotobuf=3.15.8=hcf210ce_1
libsodium=1.0.18=h1de35cc_0
libssh2=1.9.0=ha12b0ac_1
libtasn1=4.16.0=h9ed2024_0
libtiff=4.2.0=h87d7836_0
libunistring=0.9.10=h9ed2024_0
libvpx=1.11.0=he9d5cce_0
libwebp=1.2.0=hacca55c_0
libwebp-base=1.2.0=h9ed2024_0
libxml2=2.9.12=h93ec3fd_0
libzlib=1.2.11=h9173be1_1013
llvm-openmp=12.0.0=h0dcd299_1
locket=0.2.1=py38hecd8cb5_1
lz4-c=1.9.3=h23ab428_1
markdown=3.3.4=py38hecd8cb5_0
markupsafe=2.0.1=py38h9ed2024_0
matplotlib=3.5.0=py38hecd8cb5_0
matplotlib-base=3.5.0=py38h4f681db_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mightymosaic=1.2.3=pypi_0
mistune=0.8.4=py38h1de35cc_1001
multidict=5.2.0=py38hca72f7f_2
munkres=1.1.4=py_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.3.0=py38hecd8cb5_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.3=hca72f7f_2
nest-asyncio=1.5.1=pyhd3eb1b0_0
nettle=3.7.3=h230ac6f_1
networkx=2.6.3=pyhd3eb1b0_0
notebook=6.4.6=py38hecd8cb5_0
numexpr=2.8.0=py38ha53d530_1
numpy=1.22.0=py38h9d72dae_0
oauthlib=3.1.1=pyhd3eb1b0_0
olefile=0.46=pyhd3eb1b0_0
opencv=4.5.1=py38h50d1736_1
openh264=2.1.1=h8346a28_0
openssl=1.1.1n=hca72f7f_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.5=py38h743cdd8_0
pandocfilters=1.4.3=py38hecd8cb5_1
parso=0.8.3=pyhd3eb1b0_0
partd=1.2.0=pyhd3eb1b0_0
patchify=0.2.3=pypi_0
pcre=8.45=h23ab428_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.4.0=py38h98e4679_0
pip=21.2.4=py38hecd8cb5_0
pixman=0.40.0=h9ed2024_1
pluggy=1.0.0=py38hecd8cb5_0
prometheus_client=0.12.0=pyhd3eb1b0_0
promise=2.3=py38hecd8cb5_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.15.8=py38ha048514_0
psutil=5.8.0=py38h9ed2024_1
ptyprocess=0.7.0=pyhd3eb1b0_2
py=1.10.0=pyhd3eb1b0_0
py-opencv=4.5.1=py38h6c79ece_1
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pydot=1.4.2=pypi_0
pydotplus=2.0.2=pypi_0
pygments=2.10.0=pyhd3eb1b0_0
pyjwt=2.1.0=py38hecd8cb5_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=3.0.4=pyhd3eb1b0_0
pyrsistent=0.18.0=py38hca72f7f_0
pysocks=1.7.1=py38_1
pytest=6.2.5=py38hecd8cb5_2
pytest-runner=5.3.1=pyhd3eb1b0_0
python=3.8.12=h17280f6_2_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
python-flatbuffers=1.12=pyhd8ed1ab_1
python_abi=3.8=2_cp38
pytz=2021.3=pyhd3eb1b0_0
pywavelets=1.1.1=py38haf1e3a3_2
pyyaml=6.0=py38hca72f7f_1
pyzmq=22.3.0=py38he9d5cce_2
re2=2021.08.01=he49afe7_0
readline=8.1.2=hca72f7f_1
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-image=0.18.3=py38hae1ba45_0
scikit-learn=1.0.2=py38hae1ba45_0
scipy=1.8.0=py38hd329d04_1
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=58.0.4=py38hecd8cb5_0
six=1.16.0=pyhd3eb1b0_0
snappy=1.1.8=hb1e8313_0
sqlite=3.37.0=h707629a_0
svt-av1=0.8.7=he49afe7_1
tensorboard=2.4.0=pyhc547734_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.4.1=py38h50d1736_0
tensorflow-base=2.4.1=py38h8860697_0
tensorflow-datasets=1.2.0=py38_0
tensorflow-estimator=2.4.1=pyheb71bc4_0
tensorflow-examples=035fb73d5ff8b74958c9e3f83f44fc20dfc39119-=pypi_0
tensorflow-metadata=0.14.0=pyhe6710b0_1
termcolor=1.1.0=py38hecd8cb5_1
terminado=0.9.4=py38hecd8cb5_0
testpath=0.5.0=pyhd3eb1b0_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2020.10.1=py38h0cf3a3e_2
tk=8.6.11=h7bc2e8c_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.11.2=pyhd3eb1b0_0
tornado=6.1=py38h9ed2024_0
tqdm=4.62.3=pyhd3eb1b0_1
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.7=pyhd3eb1b0_0
visualkeras=0.0.2=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py38_1
werkzeug=2.0.2=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.1=py38_0
wrapt=1.13.3=py38hca72f7f_2
x264=1!161.3030=h0d85af4_1
x265=3.5=h940c156_1
xz=5.2.5=h1de35cc_0
yaml=0.2.5=haf1e3a3_0
yarl=1.6.3=py38h9ed2024_0
yellowbrick=1.3.post1=pyhd8ed1ab_1
zeromq=4.3.4=h23ab428_0
zipp=3.7.0=pyhd3eb1b0_0
zlib=1.2.11=h9173be1_1013
zstd=1.4.9=h322a384_0
